Map component: (credits) text

Map component that adds a text, typically used as credits. This function is the same as [tm_title()] but with different default values.

Usage

tm_geocoder(stack, position, group_id, z)

Arguments

stack

stack with other map components, either `"vertical"` or `"horizontal"`.

position

The position specification of the component: an object created with `tm_pos_in()` or `tm_pos_out()`. Or, as a shortcut, a vector of two values, specifying the x and y coordinates. The first is `"left"`, `"center"` or `"right"` (or upper case, meaning tighter to the map frame), the second `"top"`, `"center"` or `"bottom"`. Numeric values are also supported, where 0, 0 means left bottom and 1, 1 right top. See also vignette about positioning. In case multiple components should be combined (stacked), use `group_id` and specify `component` in [tm_comp_group()].

group_id

Component group id name. All components (e.g. legends, titles, etc) with the same `group_id` will be grouped. The specifications of how they are placed (e.g. stacking, margins etc.) are determined in [tm_comp_group()] where its argument `id` should correspond to `group_id`.

z

z index, e.g. the place of the component relative to the other componets

See also